Isolate platform-dependent comms logic
