+++ c_src/render_script.c
@@ -8,14 +8,11 @@ functions to play a compiled render script
 #include <math.h>
 #include <stdio.h>
 
-#ifdef _MSC_VER
-#include "windows_utils.h"
-#endif
+#include "comms.h"
 
 #include <GLFW/glfw3.h>
 #include <stdlib.h>
 
-#include "comms.h"
 #include "nanovg/nanovg.h"
 #include "render_script.h"
 #include "tx.h"
