+++ c_src/unix_comms.h
@@ -1,0 +1,18 @@
+/*
+# Implementations to handle port communication on UNIX-based systems
+*/
+
+#ifndef _UNIX_COMMS_H
+  #define _UNIX_COMMS_H
+
+  #include <stdio.h>
+  #include <unistd.h>
+  #include <poll.h>
+  #include <sys/time.h>
+  #include <sys/select.h>
+  #include <stdint.h>
+  #include <string.h>
+
+  #include "types.h"
+  #include "utils.h"
+#endif
